requirements.txt
# 🔌 电力知识库与AI预测平台 - Python依赖文件
# 包含Django后端、AI预测功能的所有Python依赖包

# ================================
# Django Web框架及相关组件
# ================================
Django==4.2.7                    # Django Web框架
django-cors-headers==4.3.1       # 跨域请求处理
django-mdeditor==0.1.20          # Markdown编辑器
django-ninja==1.0.1              # 现代化API框架
Pillow>=10.0.0                   # 图像处理库（兼容Python 3.13）
pycryptodome==3.19.0             # 加密库
PyJWT==2.8.0                     # JWT令牌处理

# ================================
# 数据处理
# ================================
pandas>=2.0.0                    # 数据分析库
numpy>=1.24.0                    # 数值计算库（兼容Python 3.13）

# ================================
# 机器学习
# ================================
scikit-learn>=1.3.0,<1.6.0       # 机器学习库（稳定版本）
xgboost>=2.0.0                   # 梯度提升算法（支持Python 3.13）
joblib>=1.3.0                    # 模型序列化

# ================================
# 数据可视化
# ================================
plotly>=5.17.0                   # 交互式图表库

# ================================
# 可选深度学习支持（按需启用）
# ================================
# tensorflow>=2.13.0             # 深度学习框架
# keras>=2.13.0                  # 高级神经网络API

# ================================
# 数据库支持（按需选择）
# ================================
# mysqlclient==2.2.0             # MySQL数据库连接器
# psycopg2-binary==2.9.0         # PostgreSQL数据库连接器

# ================================
# 开发调试工具（可选）
# ================================
# python-dotenv>=1.0.0           # 环境变量管理
# django-extensions>=3.2.0       # Django扩展工具
# django-debug-toolbar>=4.0.0    # 调试工具栏

# ================================
# 大模型知识库（RAG）相关依赖
# ================================
PyPDF2>=3.0.0                    # PDF文档处理
python-docx>=0.8.11              # Word文档处理
markdown>=3.5.0                  # Markdown处理
beautifulsoup4>=4.12.0           # HTML解析
langchain>=0.1.0                 # LangChain框架
jieba>=0.42.1                    # 中文分词
openai>=1.0.0                    # OpenAI API支持
anthropic>=0.8.0                 # Anthropic API支持
zhipuai>=2.0.0                   # 智谱AI API支持
requests>=2.31.0                 # HTTP请求库
